refactor: Move DefaultTransitGroupCalculator to grouppriority package
