+++ src/test/java/org/opentripplanner/transit/model/network/grouppriority/DefaultTransitGroupCalculatorTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptoradapter.transit.cost;
+package org.opentripplanner.transit.model.network.grouppriority;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
